requirements.txt
feedparser
discord.py